+++ changelog.md
@@ -12,6 +12,7 @@ __[changes]__
 __[maintenance]__
 
 * NLog library is updated from 6.0.7 to 6.1.0.
+* Update certificate information for Blender LTS installer.
 
 ## Version 2026.01.31.0